backend/requirements.txt
asgiref==3.2.7
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
PyJWT==1.7.1
pytz==2020.1
sqlparse==0.3.1
